restore gssapi.py file
