+++ impacket/krb5/gssapi.py
@@ -16,7 +16,6 @@
 #
 # Author:
 #   Alberto Solino (@agsolino)
-#   Thomas Seigneuret (@zblurx)
 #
 import struct
 import random
@@ -50,9 +49,6 @@ GSS_HMAC = 0x11
 # Wrap Semantics
 GSS_RC4  = 0x10
 
-# Kerberos OID: 1.2.840.113554.1.2.2
-KRB_OID = b'\x06\t*\x86H\x86\xf7\x12\x01\x02\x02'
-
 # 2.  Key Derivation for Per-Message Tokens
 KG_USAGE_ACCEPTOR_SEAL  = 22
 KG_USAGE_ACCEPTOR_SIGN  = 23
@@ -60,6 +56,9 @@ KG_USAGE_INITIATOR_SEAL = 24
 KG_USAGE_INITIATOR_SIGN = 25
 
 KRB5_AP_REQ = struct.pack('<H', 0x1)
+
+# Kerberos OID: 1.2.840.113554.1.2.2
+KRB_OID = b'\x06\t*\x86H\x86\xf7\x12\x01\x02\x02'
 
 class MechIndepToken():
     
@@ -168,7 +167,78 @@ class GSSAPI_RC4:
         finalData = GSS_GETMIC_HEADER + token.getData()
         return finalData
    
-    def GSS_Wrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True):
+    def GSS_Wrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True, authData=None):
+        # Damn inacurate RFC, useful info from here
+        # https://social.msdn.microsoft.com/Forums/en-US/fb98e8f4-e697-4652-bcb7-604e027e14cc/gsswrap-token-size-kerberos-and-rc4hmac?forum=os_windowsprotocols
+        # and here
+        # http://www.rfc-editor.org/errata_search.php?rfc=4757
+        GSS_WRAP_HEADER = b'\x60\x2b\x06\x09\x2a\x86\x48\x86\xf7\x12\x01\x02\x02'
+        token = self.WRAP()
+
+        # Let's pad the data
+        pad = (8 - (len(data) % 8)) & 0x7
+        padStr = b(chr(pad)) * pad
+        data += padStr
+
+        token['SGN_ALG'] = GSS_HMAC
+        token['SEAL_ALG'] = GSS_RC4
+
+        if direction == 'init':
+            token['SND_SEQ'] = struct.pack('>L', sequenceNumber) + b'\x00'*4
+        else:
+            token['SND_SEQ'] = struct.pack('>L', sequenceNumber) + b'\xff'*4
+
+        # Random confounder :)
+        token['Confounder'] = b(''.join([rand.choice(string.ascii_letters) for _ in range(8)]))
+
+        Ksign = HMAC.new(sessionKey.contents, b'signaturekey\0', MD5).digest()
+        Sgn_Cksum = MD5.new(struct.pack('<L',13) + token.getData()[:8] + token['Confounder'] + data).digest()
+
+        Klocal = bytearray()
+        from builtins import bytes
+        for n in bytes(sessionKey.contents):
+            Klocal.append( n ^ 0xF0)
+
+        Kcrypt = HMAC.new(Klocal,struct.pack('<L',0), MD5).digest()
+        Kcrypt = HMAC.new(Kcrypt,struct.pack('>L', sequenceNumber), MD5).digest()
+        
+        Sgn_Cksum = HMAC.new(Ksign, Sgn_Cksum, MD5).digest()
+
+        token['SGN_CKSUM'] = Sgn_Cksum[:8]
+
+        Kseq = HMAC.new(sessionKey.contents, struct.pack('<L',0), MD5).digest()
+        Kseq = HMAC.new(Kseq, token['SGN_CKSUM'], MD5).digest()
+
+        token['SND_SEQ'] = ARC4.new(Kseq).encrypt(token['SND_SEQ'])
+
+        if authData is not None:
+            from impacket.dcerpc.v5.rpcrt import SEC_TRAILER
+            wrap = self.WRAP(authData[len(SEC_TRAILER()) + len(GSS_WRAP_HEADER):])
+            snd_seq = wrap['SND_SEQ']
+
+            Kseq = HMAC.new(sessionKey.contents, struct.pack('<L',0), MD5).digest()
+            Kseq = HMAC.new(Kseq, wrap['SGN_CKSUM'], MD5).digest()
+
+            snd_seq = ARC4.new(Kseq).encrypt(wrap['SND_SEQ'])
+ 
+            Kcrypt = HMAC.new(Klocal,struct.pack('<L',0), MD5).digest()
+            Kcrypt = HMAC.new(Kcrypt,snd_seq[:4], MD5).digest()
+            rc4 = ARC4.new(Kcrypt)
+            cipherText = rc4.decrypt(token['Confounder'] + data)[8:]
+        elif encrypt is True:
+            rc4 = ARC4.new(Kcrypt)
+            token['Confounder'] = rc4.encrypt(token['Confounder'])
+            cipherText = rc4.encrypt(data)
+        else:
+            cipherText = data
+
+        finalData = GSS_WRAP_HEADER + token.getData()
+        return cipherText, finalData
+
+    def GSS_Unwrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True, authData=None):
+        return self.GSS_Wrap(sessionKey, data, sequenceNumber, direction, encrypt, authData)
+
+    def GSS_Wrap_LDAP(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True):
        
         if encrypt:
             data += b'\x01'
@@ -229,7 +299,7 @@ class GSSAPI_RC4:
             finalData, cipherText = MechIndepToken(token.getData() + cipherText, KRB_OID).to_bytes()
             return cipherText, finalData
 
-    def GSS_Unwrap(self, sessionKey, data, sequenceNumber, direction = 'init'):
+    def GSS_Unwrap_LDAP(self, sessionKey, data, sequenceNumber, direction = 'init'):
         return self.GSS_Wrap(sessionKey, data, sequenceNumber, direction, encrypt=False)
 
 class GSSAPI_AES():
@@ -284,7 +354,50 @@ class GSSAPI_AES():
         result = data[numBytes:] + data[:numBytes]
         return result
         
-    def GSS_Wrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True, use_padding=False):
+    def GSS_Wrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True):
+        token = self.WRAP()
+
+        cipher = self.cipherType()
+
+        # Let's pad the data
+        pad = (cipher.blocksize - (len(data) % cipher.blocksize)) & 15
+        padStr = b'\xFF' * pad
+        data += padStr
+
+        # The RRC field ([RFC4121] section 4.2.5) is 12 if no encryption is requested or 28 if encryption 
+        # is requested. The RRC field is chosen such that all the data can be encrypted in place.
+        rrc = 28
+
+        token['Flags'] = 6
+        token['EC'] = pad
+        token['RRC'] = 0
+        token['SND_SEQ'] = struct.pack('>Q',sequenceNumber)
+
+        cipherText = cipher.encrypt(sessionKey, KG_USAGE_INITIATOR_SEAL,  data + token.getData(), None)
+        token['RRC'] = rrc
+
+        cipherText = self.rotate(cipherText, token['RRC'] + token['EC'])
+
+        #nn = self.unrotate(cipherText, token['RRC'] + token['EC'])
+        ret1 = cipherText[len(self.WRAP()) + token['RRC'] + token['EC']:]
+        ret2 = token.getData() + cipherText[:len(self.WRAP()) + token['RRC'] + token['EC']]
+
+        return ret1, ret2
+
+    def GSS_Unwrap(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True, authData=None):
+        from impacket.dcerpc.v5.rpcrt import SEC_TRAILER
+
+        cipher = self.cipherType()
+        token = self.WRAP(authData[len(SEC_TRAILER()):])
+
+        rotated = authData[len(self.WRAP())+len(SEC_TRAILER()):] + data
+ 
+        cipherText = self.unrotate(rotated, token['RRC'] + token['EC'])
+        plainText = cipher.decrypt(sessionKey, KG_USAGE_ACCEPTOR_SEAL,  cipherText)
+
+        return plainText[:-(token['EC']+len(self.WRAP()))], None
+
+    def GSS_Wrap_LDAP(self, sessionKey, data, sequenceNumber, direction = 'init', encrypt=True, use_padding=False):
         token = self.WRAP()
 
         cipher = self.cipherType()
@@ -318,7 +431,7 @@ class GSSAPI_AES():
 
         return ret1, ret2
 
-    def GSS_Unwrap(self, sessionKey, data, sequenceNumber, direction = 'init'):
+    def GSS_Unwrap_LDAP(self, sessionKey, data, sequenceNumber, direction = 'init'):
 
         cipher = self.cipherType()
         token = self.WRAP(data[:16])
